==> 2 Analysis/Phosphoproteome/Independent/1p Single normalising data.Rmd <==
---
title: "Sample loading normalisation + IRS for phosphoproteomics"
output:
---

```{r}
#######################################################################################################################################
# Independent analysis Notebook 1 - DESCRIPTION
#
# Here we take the raw data (output of Perseus without NA-filtering).
# First, we infer data for the missing sample - fortunately this was one of the middle samples, so data is inferred by taking the 
# mean of the adjacent time points.
# This is the "independent" analysis - i.e. analysing WT and CKO independently.
# Therefore we split the dataset into WT + CKO before doing the normalisation steps to each.
# We then apply sample loading normalisation to correct for small errors in pipetting - this assumes equal protein was used for
# all samples submitted for TMT labelling.
# We then apply Internal Reference Scaling as described by Phil Wilmarth - https://github.com/pwilmart/IRS_normalization
# This corrects for batch effects.
# Along the way we plot the data to make sure that we can spot obvious mistakes.
#
# The extra step taken here for the phosphoproteome is correcting for relative changes in the amount of the associated protein.
# This is because we want to find out changes in phosphorylation, rather than simply changes in the amount of protein that happens to be phosphorylated.
# For this, we need the normalised data from the proteomics (independent) analysis.
# Hence all the phosphopeptides included in subsequent analyses have associated proteins detected in our proteomics analysis.
#
# Finally, we export the data because this is handy for plotting proteins of interest in other software!
#
# INPUTS = raw data from Perseus (.txt), all normalised data from independent proteomics analysis (.csv)
# OUTPUTS = normalised data for further analysis (.csv)
#
#######################################################################################################################################
```

```{r}
# Setup

# Clear your workspace
rm(list=ls())

# Load useful packages
if (!require("tidyverse")) {
  install.packages("tidyverse", dependencies = TRUE)
  library(tidyverse)
}

if (!require("BiocManager")) {
  install.packages("BiocManager", dependencies = TRUE)
}

if (!require("edgeR")) {
  BiocManager::install("edgeR")
  library(edgeR)
} 

if (!require("psych")) {
  install.packages("psych", dependencies = TRUE)
  library(psych)
} 
# Make a list of all the .txt files in your working directory just so you can see what's going on
X = list.files(".",".txt")

# read the raw data file, and ensure "NaN" is interpreted as NA
data_start <- read.delim("CRM3962_6set_MQ1633C_phos_noFil_m8b.txt", na = c("NaN"), stringsAsFactors = F)
data_start <- data_start[-1,] # Remove the first row which is just junk
data_start[, 1:60] <- sapply(data_start[,1:60], as.numeric) # Make the data numeric
missing_data <- apply(cbind(data_start[,13], data_start[,15]), 1, mean) # Infer missing values for CKO sample 12
data_start <- cbind(data_start[1:13], missing_data, data_start[15:ncol(data_start)])

```

```{r}
# Create column containing protein_phosphosite_multiplicity

data_start$Phosphosite <- paste(data_start$Protein, "_", data_start$Amino.acid, data_start$Position, data_start$Multiplicity, sep = "")
  
# Selecting the main gene name from each line
genes <- vector("list", length(data_start$Gene.names)) # Create empty vector
for (i in 1:length(data_start$Gene.names)) {
  genes[i] <- strsplit(data_start$Gene.names[i], ";")
}
gene_names <- unlist(lapply(genes, '[', 1))
data_start$main_gene <- gene_names
```

```{r}
# Separate the data into WT + CKO
WT_data_start <- cbind(data_start[,31:60], data_start[,61:ncol(data_start)])
CKO_data_start <- cbind(data_start[,1:30], data_start[,61:ncol(data_start)])

# Remove rows which contain NaN within the data columns

WT_data_start <- WT_data_start[rowSums(is.na(WT_data_start[, 1:30])) == 0,]
CKO_data_start <- CKO_data_start[rowSums(is.na(CKO_data_start[, 1:30])) == 0,]

# Extract data only
WT_data_raw <- as.data.frame(WT_data_start[1:30])
CKO_data_raw <- as.data.frame(CKO_data_start[1:30])

```

```{r}
#See what raw data looks like 

boxplot(log2(WT_data_raw), col = rep(c("red", "green", "blue"), each = 10), 
        main = "WT Raw data: \nExp1 (red), Exp2 (green), Exp3 (blue)",
        xlab = 'TMT Sample', ylab = 'log2 of Intensity', notch = TRUE)

boxplot(log2(CKO_data_raw), col = rep(c("red", "green", "blue"), each = 10), 
        main = "CKO Raw data: \nExp1 (red), Exp2 (green), Exp3 (blue)",
        xlab = 'TMT Sample', ylab = 'log2 of Intensity', notch = TRUE)

# can also look at density plots (like a distribution histogram)
plotDensities(log2(WT_data_raw), col = rep(c('red', 'green', 'blue'), 6), 
              main = 'WT Raw data')

plotDensities(log2(CKO_data_raw), col = rep(c('red', 'green', 'blue'), 6), 
              main = 'CKO Raw data')

# check the column totals (per channel sums)
format(round(colSums(WT_data_raw), digits = 0), big.mark = ",")
format(round(colSums(CKO_data_raw), digits = 0), big.mark = ",")
```

```{r}
#Sample loading normalisation + IRS Normalisation (Phil Wilmarth, OHSU PSR Core, January 2018)

# separate the TMT data by experiment
exp1_raw <- WT_data_raw[c(1:10)]
exp2_raw <- WT_data_raw[c(11:20)]
exp3_raw <- WT_data_raw[c(21:30)]
exp4_raw <- CKO_data_raw[c(1:10)]
exp5_raw <- CKO_data_raw[c(11:20)]
exp6_raw <- CKO_data_raw[c(21:30)]

# figure out the global scaling value
WT_target <- mean(c(colSums(exp1_raw), colSums(exp2_raw), colSums(exp3_raw)))
CKO_target <- mean(c(colSums(exp4_raw), colSums(exp5_raw), colSums(exp6_raw)))

# do the sample loading normalization before the IRS normalization
# there is a different correction factor for each column
norm_facs <- WT_target / colSums(exp1_raw)
exp1_sl <- sweep(exp1_raw, 2, norm_facs, FUN = "*")
norm_facs <- WT_target / colSums(exp2_raw)
exp2_sl <- sweep(exp2_raw, 2, norm_facs, FUN = "*")
norm_facs <- WT_target / colSums(exp3_raw)
exp3_sl <- sweep(exp3_raw, 2, norm_facs, FUN = "*")
norm_facs <- CKO_target / colSums(exp4_raw)
exp4_sl <- sweep(exp4_raw, 2, norm_facs, FUN = "*")
norm_facs <- CKO_target / colSums(exp5_raw)
exp5_sl <- sweep(exp5_raw, 2, norm_facs, FUN = "*")
norm_facs <- CKO_target / colSums(exp6_raw)
exp6_sl <- sweep(exp6_raw, 2, norm_facs, FUN = "*")

# make a pre-IRS data frame after sample loading norms
WT_data_sl <- cbind(exp1_sl, exp2_sl, exp3_sl)
CKO_data_sl <- cbind(exp4_sl, exp5_sl, exp6_sl)

# see what the SL normalised data look like
boxplot(log2(WT_data_sl), col = rep(c("red", "green", "blue"), each = 10), 
        main = "WT Sample loading (SL) normalized data: \nExp1 (red), Exp2 (green), Exp3 (blue)",
        xlab = 'TMT Sample', ylab = 'log2 of Intensity', notch = TRUE)
boxplot(log2(CKO_data_sl), col = rep(c("red", "green", "blue"), each = 10), 
        main = "CKO Sample loading (SL) normalized data: \nExp1 (red), Exp2 (green), Exp3 (blue)",
        xlab = 'TMT Sample', ylab = 'log2 of Intensity', notch = TRUE)

# can also look at density plots (like a distribution histogram)    
plotDensities(log2(WT_data_sl), col = rep(c("red", "green", "blue"), 10), main = "SL data")
plotDensities(log2(CKO_data_sl), col = rep(c("red", "green", "blue"), 10), main = "SL data")

# check the columnn totals
format(round(colSums(WT_data_sl), digits = 0), big.mark = ",")
format(round(colSums(CKO_data_sl), digits = 0), big.mark = ",")
```

```{r}
# make working frame with row sums from each frame
WT_irs <- tibble(apply(exp1_sl[,9:10], 1, mean), apply(exp2_sl[,9:10], 1, mean), apply(exp3_sl[,9:10], 1, mean))
CKO_irs <- tibble(apply(exp4_sl[,9:10], 1, mean), apply(exp5_sl[,9:10], 1, mean), apply(exp6_sl[,9:10], 1, mean))

colnames(WT_irs) <- c("mean1", "mean2", "mean3")
colnames(CKO_irs) <- c("mean4", "mean5", "mean6")

# get the average intensity for each protein
WT_irs$average <- apply(WT_irs, 1, mean)
CKO_irs$average <- apply(CKO_irs, 1, mean)

# compute the scaling factor vectors
WT_irs$fac1 <- WT_irs$average / WT_irs$mean1
WT_irs$fac2 <- WT_irs$average / WT_irs$mean2
WT_irs$fac3 <- WT_irs$average / WT_irs$mean3
CKO_irs$fac4 <- CKO_irs$average / CKO_irs$mean4
CKO_irs$fac5 <- CKO_irs$average / CKO_irs$mean5
CKO_irs$fac6 <- CKO_irs$average / CKO_irs$mean6

# make new data frame with normalized data
WT_data_irs <- exp1_sl * WT_irs$fac1
WT_data_irs <- cbind(WT_data_irs, exp2_sl * WT_irs$fac2)
WT_data_irs <- cbind(WT_data_irs, exp3_sl * WT_irs$fac3)

CKO_data_irs <- exp4_sl * CKO_irs$fac4
CKO_data_irs <- cbind(CKO_data_irs, exp5_sl * CKO_irs$fac5)
CKO_data_irs <- cbind(CKO_data_irs, exp6_sl * CKO_irs$fac6)

# see what the IRS data look like
boxplot(log2(WT_data_irs), col = rep(c("red", "green", "blue"), each = 10), 
        main = "WT Internal Reference Scaling (IRS) normalized data: \nExp1 (red), Exp2 (green), Exp3 (blue)",
        xlab = 'TMT Sample', ylab = 'log2 of Intensity', notch = TRUE)
boxplot(log2(CKO_data_irs), col = rep(c("red", "green", "blue"), each = 10), 
        main = "CKO Internal Reference Scaling (IRS) normalized data: \nExp1 (red), Exp2 (green), Exp3 (blue)",
        xlab = 'TMT Sample', ylab = 'log2 of Intensity', notch = TRUE)

# can also look at density plots (like a distribution histogram)    
plotDensities(log2(WT_data_irs), col = rep(c("red", "green", "blue"), 10), main = "IRS data")
plotDensities(log2(CKO_data_irs), col = rep(c("red", "green", "blue"), 10), main = "IRS data")

# check column totals
format(round(colSums(WT_data_irs), digits = 0), big.mark = ",")
format(round(colSums(CKO_data_irs), digits = 0), big.mark = ",")

```

```{r}
# Load the list of total proteome normalised data
WT_total_proteome <- read.csv("All_WT_normalised_data.csv")
CKO_total_proteome <- read.csv("All_CKO_normalised_data.csv")
# Data only
WT_proteome_data <- WT_total_proteome[, -1:-3]
CKO_proteome_data <- CKO_total_proteome[, -1:-3]

# Normalise all values according to the protein abundance (from total proteome analysis)
WT_total_proteome_max <- as.numeric(apply(WT_proteome_data, 1, max))
# Initialise matrix
WT_total_proteome_normalised <- matrix(0, nrow(WT_proteome_data), ncol(WT_proteome_data))
# Divide protein intensities by their maximum
for (i in 1:nrow(WT_proteome_data)) {
  WT_total_proteome_normalised[i,] <- t(apply(WT_proteome_data[i, ], 1, function(x) x / WT_total_proteome_max[i] ))
}
# Make it a data frame and attach protein + gene names
WT_total_proteome_normalised <- as.data.frame(WT_total_proteome_normalised)
WT_total_proteome_normalised <- cbind(WT_total_proteome[, 1:3], WT_total_proteome_normalised)

# Select only phosphopeptides whose proteins are present in the total proteome
WT_data_irs <- WT_data_irs %>% dplyr::select(-contains("P")) # Remove the pools from the dataframe
WT_data_irs_adj <- cbind(WT_data_irs, WT_data_start[,31:ncol(WT_data_start)])
WT_data_irs_adj <- WT_data_irs_adj[which(WT_data_irs_adj$Protein %in% WT_total_proteome_normalised$uniprot),]

# Normalise the phospho data
WT_phospho_normalised <- as.data.frame(matrix(0, nrow = nrow(WT_data_irs_adj), ncol(WT_data_irs))) # initialise matrix
for (i in 1:nrow(WT_data_irs_adj)) {
  WT_phospho_normalised[i,] <- unlist(apply(WT_data_irs_adj[i, 1:24], 1, 
                                    function(x) x / WT_total_proteome_normalised[which(WT_total_proteome_normalised$uniprot %in% WT_data_irs_adj$Protein[i]),4:ncol(WT_total_proteome_normalised)]))
}
WT_phospho_normalised <- cbind(WT_data_irs_adj$Phosphosite, 
                               WT_data_irs_adj$Protein, 
                               WT_data_irs_adj$Protein.names, 
                               WT_data_irs_adj$main_gene,
                               WT_data_irs_adj$Sequence.window,
                               WT_phospho_normalised)
colnames(WT_phospho_normalised)[1:5] <- c("Phosphosite", "Protein", "Protein names", "main_gene", "Sequence.window")




# Same thing for CKO

CKO_total_proteome_max <- as.numeric(apply(CKO_proteome_data, 1, max))
# Initialise matrix
CKO_total_proteome_normalised <- matrix(0, nrow(CKO_proteome_data), ncol(CKO_proteome_data))
# Divide protein intensities by their maximum
for (i in 1:nrow(CKO_proteome_data)) {
  CKO_total_proteome_normalised[i,] <- t(apply(CKO_proteome_data[i, ], 1, function(x) x / CKO_total_proteome_max[i] ))
}
# Make it a data frame and attach protein + gene names
CKO_total_proteome_normalised <- as.data.frame(CKO_total_proteome_normalised)
CKO_total_proteome_normalised <- cbind(CKO_total_proteome[, 1:3], CKO_total_proteome_normalised)

# Select only phosphopeptides whose proteins are present in the total proteome
CKO_data_irs <- CKO_data_irs %>% dplyr::select(-contains("P")) # Remove the pools from the dataframe
CKO_data_irs_adj <- cbind(CKO_data_irs, CKO_data_start[,31:ncol(CKO_data_start)])
CKO_data_irs_adj <- CKO_data_irs_adj[which(CKO_data_irs_adj$Protein %in% CKO_total_proteome_normalised$uniprot),]

# Normalise the phospho data
CKO_phospho_normalised <- as.data.frame(matrix(0, nrow = nrow(CKO_data_irs_adj), ncol(CKO_data_irs))) # initialise matrix
for (i in 1:nrow(CKO_data_irs_adj)) {
  CKO_phospho_normalised[i,] <- unlist(apply(CKO_data_irs_adj[i, 1:24], 1, 
                                    function(x) x / CKO_total_proteome_normalised[which(CKO_total_proteome_normalised$uniprot %in% CKO_data_irs_adj$Protein[i]),4:ncol(CKO_total_proteome_normalised)]))
}
CKO_phospho_normalised <- cbind(CKO_data_irs_adj$Phosphosite, 
                                CKO_data_irs_adj$Protein, 
                                CKO_data_irs_adj$Protein.names, 
                                CKO_data_irs_adj$main_gene,
                                CKO_data_irs_adj$Sequence.window, 
                                CKO_phospho_normalised)
colnames(CKO_phospho_normalised)[1:5] <- c("Phosphosite", "Protein", "Protein names", "main_gene", "Sequence.window")

```


```{r}
# see how things cluster now that we have nice boxplots and density plots
plotMDS(log2(WT_phospho_normalised[,6:ncol(WT_phospho_normalised)]), col = rep(c("red", "green", "blue"), each = 10), 
        main = "SL/TMM clusters group by TMT experiment")
plotMDS(log2(CKO_phospho_normalised[,6:ncol(CKO_phospho_normalised)]), col = rep(c("red", "green", "blue"), each = 10), 
        main = "SL/TMM clusters group by TMT experiment")
```


```{r}
# Reorganise the dataframes to export a nice spreadsheet

write.csv(WT_phospho_normalised, "WT_phospho_normalised.csv", row.names = F)
write.csv(CKO_phospho_normalised, "CKO_phospho_normalised.csv", row.names = F)

```

```{r}
# Prepare .txt files for eJTK

# Separate into WT + CKO with only uniprot IDs
WTeJTKcomb <- cbind(WT_phospho_normalised$Phosphosite, WT_phospho_normalised[,6:ncol(WT_phospho_normalised)])

CKOeJTKcomb <- cbind(CKO_phospho_normalised$Phosphosite, CKO_phospho_normalised[,6:ncol(CKO_phospho_normalised)])

# Colnames need to be: "#" for IDs, then CT or ZT preceding timepoint number
colnames(WTeJTKcomb) <- c("ID", paste(rep("CT", 24), seq(0, 69, 3), sep=""))
colnames(CKOeJTKcomb) <- c("ID", paste(rep("CT", 24), seq(0, 69, 3), sep=""))

# Files need to be tab-separated
write.table(WTeJTKcomb, file = "WTeJTKcomb-single-p.txt", sep = "\t", row.names = F, col.names = T, quote = F)
write.table(CKOeJTKcomb, file = "CKOeJTKcomb-single-p.txt", sep = "\t", row.names = F, col.names = T, quote = F)

```
